+++ packages/contracts/contracts/plugin/IPluginSetup.sol
@@ -21,7 +21,7 @@ interface IPluginSetup {
         returns (
             address plugin,
             address[] memory helpers,
-            PermissionLib.ItemMultiTarget[] memory permissions
+            PermissionLib.MultiTargetPermission[] memory permissions
         );
 
     /// @notice The ABI required to decode the `bytes` data in `prepareUpdate()`.
@@ -51,7 +51,7 @@ interface IPluginSetup {
         returns (
             address[] memory updatedHelpers,
             bytes memory initData,
-            PermissionLib.ItemMultiTarget[] memory permissions
+            PermissionLib.MultiTargetPermission[] memory permissions
         );
 
     /// @notice The ABI required to decode the `bytes` data in `prepareUninstallation()`.
@@ -70,7 +70,7 @@ interface IPluginSetup {
         address _plugin,
         address[] calldata _currentHelpers,
         bytes calldata _data
-    ) external returns (PermissionLib.ItemMultiTarget[] memory permissions);
+    ) external returns (PermissionLib.MultiTargetPermission[] memory permissions);
 
     /// @notice Returns the plugin's base implementation.
     /// @return address The address of the plugin implementation contract.